Allow file renames to change directories
I was prohibiting a user from entering any '/' in any new file name but
I now allow it so that a user can move a file to a different directory.
It's crude and there are probably some dangerous name change potentials
here now but it is better to allow flexibility.

User can move a file up a directory by renaming to "../file.rec".